feature update

+++ src/Core/ProductApp.Application/Common/ProductMessage.cs
@@ -5,6 +5,6 @@ public class ProductMessage
     public long Id { get; set; }
     public string? Name { get; set; }
     public decimal? Value { get; set; }
-    public int? Quantity { get; set; }
+    public long? Quantity { get; set; }
     public short Status { get; set; }
 }
